requirements.txt
# ============================================
# 微信AI助手项目依赖
# ============================================

# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# HTTP客户端
httpx==0.25.1
aiohttp==3.9.1

# 微信SDK（可选，用于便捷处理）
# wechatpy==1.8.18

# 数据验证
pydantic==2.5.0
pydantic-settings==2.1.0

# 日志
loguru==0.7.2

# 加密解密
cryptography==41.0.7

# 智谱AI官方SDK
zai-sdk==0.2.2

# 环境变量
python-dotenv==1.0.0

# 异步支持
asyncio==3.4.3

# 工具类
python-dateutil==2.8.2

# 数据库（后续扩展）
# sqlalchemy==2.0.23
# aiosqlite==0.19.0
